+++ src/main/java/com/book/store/Repository/BookRepository.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
    Optional<Book>  findByTitle(String title);
     @Query("SELECT b.description FROM Book b WHERE b.id = :id")
-   String  getDescriptionById(Integer id);
+   String  getDescriptionById(Long id);
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    Optional<Book> findById(Integer id);
+    Optional<Book> findById(Long id);
 }
